Impose maximum level restriction for BatchedTensors
We support 64 total levels. This is done so that we can represent lists
of levels as a bitset that fits into a single `int64_t` and is a
reasonable upper bound because we only support (physical) tensors of up
to 64 dimensions with vmap (see kVmapMaxTensorDims).

Test Plan:
`./build/bin/vmap_test`. One day we'll test this with the vmap Python
API.

ghstack-source-id: 26df902cd7a1ddd57438702f2adbd70625d183a6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/39580